Modularize application using Guice
