+++ app/src/main/java/ceres/App.java
@@ -3,11 +3,14 @@
  */
 package ceres;
 
-import ceres.configuration.ApplicationConfiguration;
+import ceres.crawler.PoemFetcher;
+import com.google.inject.Guice;
 import java.io.IOException;
 
 public class App {
     public static void main(String[] args) throws IOException {
-        var config = ApplicationConfiguration.loadFrom("app.yaml");
+        var injector = Guice.createInjector(new AppModule());
+        var poemFetcher = injector.getInstance(PoemFetcher.class);
+        poemFetcher.fetchPoems();
     }
 }
